Map 1 and 2 worked on, bugs found, baked scenes attempted
Map 1 and 2 are very similiar so i added more objects to the maps so that the player cant escape as easily. I tried baking the lights on my third scene to get some nicer lighting but it still has to be worked on. A bug was found were the checkpoint system doesn't keep track of what colliders it has collided with thus allowing the player to get checkpoint progress from the same checkpoint several times.